{"schema":"decodex/commit/1","summary":"Modularize app-server protocol surfaces","authority":"XY-1115"}
